# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Car Management API",
    "version": "1.0.0",
    "description": "API to manage Binar car data"
  },
...
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local server"
    }
  ],
  "paths": {
...
    "/user/current": {
      "get": {
        "tags": ["Users"],
        "summary": "Get current logged-in user information",
        "operationId": "getCurrentUser",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved the current user information",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "isSuccess": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "integer"
                            },
                            "name": {
                              "type": "string"
                            },
                            "age": {
                              "type": "integer"
                            },
                            "address": {
                              "type": "string"
                            },
                            "role": {
                              "type": "string"
                            },
                            "createdAt": {
                              "type": "string",
                              "format": "date-time"
                            },
                            "updatedAt": {
                              "type": "string",
                              "format": "date-time"
                            }
                          }
                        }
                      }
                    }
                  }
                },
                "examples": {
                  "example1": {
                    "summary": "A successful response example for retrieving current user info",
                    "value": {
                      "status": "Success",
                      "message": "this is the account currently logged in",
                      "isSuccess": true,
                      "data": {
                        "user": {
                          "id": 14,
                          "name": "Lela Lynch",
                          "age": 27,
                          "address": "8226 Timmothy Harbors",
                          "role": "superAdmin",
                          "createdAt": "2024-11-02T00:02:49.852Z",
                          "updatedAt": "2024-11-02T00:02:49.852Z"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failed to get cars data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Fail"
                    },
                    "message": {
                      "type": "string",
                      "example": "Internal server error"
                    },
                    "isSuccess": {
                      "type": "boolean",
                      "example": false
                    },
                    "data": {
                      "example": null
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/user/{id}": {
      "get": {
        "tags": ["Users"],
        "summary": "Get a user by ID",
        "operationId": "getUserById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the user to retrieve",
            "schema": {
              "type": "integer",
              "example": 12
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved user data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "isSuccess": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "integer"
                            },
                            "name": {
                              "type": "string"
                            },
                            "age": {
                              "type": "integer"
                            },
                            "address": {
                              "type": "string"
                            },
                            "role": {
                              "type": "string"
                            },
                            "createdAt": {
                              "type": "string",
                              "format": "date-time"
                            },
                            "updatedAt": {
                              "type": "string",
                              "format": "date-time"
                            }
                          }
                        }
                      }
                    }
                  }
                },
                "examples": {
                  "example1": {
                    "summary": "A successful response example for retrieving a user by ID",
                    "value": {
                      "status": "Success",
                      "message": "Success get user data",
                      "isSuccess": true,
                      "data": {
                        "user": {
                          "id": 12,
                          "name": "Neil Hauck",
                          "age": 29,
                          "address": "77918 Janis Expressway",
                          "role": "superAdmin",
                          "createdAt": "2024-11-02T00:02:49.849Z",
                          "updatedAt": "2024-11-02T00:02:49.849Z"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failed to get cars data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Fail"
                    },
                    "message": {
                      "type": "string",
                      "example": "Internal server error"
                    },
                    "isSuccess": {
                      "type": "boolean",
                      "example": false
                    },
                    "data": {
                      "example": null
                    }
                  }
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["Users"],
        "summary": "Update a user by ID",
        "operationId": "updateUserById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the user to update",
            "schema": {
              "type": "integer",
              "example": 12
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "age": {
                    "type": "integer"
                  },
                  "address": {
                    "type": "string"
                  }
                },
                "required": ["name", "age", "address"]
              },
              "examples": {
                "example1": {
                  "summary": "An example of updating user information",
                  "value": {
                    "name": "aqsa",
                    "age": 20,
                    "address": "suka maju"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully updated user data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "isSuccess": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "name": {
                              "type": "string"
                            },
                            "age": {
                              "type": "integer"
                            },
                            "address": {
                              "type": "string"
                            }
                          }
                        }
                      }
                    }
                  }
                },
                "examples": {
                  "example1": {
                    "summary": "A successful response example for updating a user",
                    "value": {
                      "status": "Success",
                      "message": "Success update user",
                      "isSuccess": true,
                      "data": {
                        "user": {
                          "id": "12",
                          "name": "aqsa",
                          "age": 20,
                          "address": "suka maju"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Fail"
                    },
                    "message": {
                      "type": "string",
                      "example": "Data not found"
                    },
                    "isSuccess": {
                      "type": "boolean",
                      "example": false
                    },
                    "data": {
                      "type": "array",
                      "example": null
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failed to get cars data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Fail"
                    },
                    "message": {
                      "type": "string",
                      "example": "Internal server error"
                    },
                    "isSuccess": {
                      "type": "boolean",
                      "example": false
                    },
                    "data": {
                      "example": null
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Users"],
        "summary": "Delete a user by ID",
        "operationId": "deleteUserById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the user to delete",
            "schema": {
              "type": "integer",
              "example": 12
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully deleted user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "isSuccess": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array"
                    }
                  }
                },
                "examples": {
                  "example1": {
                    "summary": "A successful response example for deleting a user",
                    "value": {
                      "status": "Success",
                      "message": "Success delete user",
                      "isSuccess": true,
                      "data": null
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Fail"
                    },
                    "message": {
                      "type": "string",
                      "example": "Data not found"
                    },
                    "isSuccess": {
                      "type": "boolean",
                      "example": false
                    },
                    "data": {
                      "type": "array",
                      "example": null
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failed to get cars data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Fail"
                    },
                    "message": {
                      "type": "string",
                      "example": "Internal server error"
                    },
                    "isSuccess": {
                      "type": "boolean",
                      "example": false
                    },
                    "data": {
                      "example": null
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}